Fix error in estimating price

+++ RideGo/app/src/main/java/com/example/ridego/data/model/RouteModels.kt
@@ -60,5 +60,5 @@ data class GeometryData(
 )
 data class ValueText(
     val text: String,
-    val value: Int
+    val value: Double // GraphHopper returns floating meters/seconds; use Double to avoid parse errors
 )
